+++ checkpoint/orbax/checkpoint/checkpoint_utils.py
@@ -13,14 +13,14 @@
 # limitations under the License.
 
 """High-level checkpoint utils provided for user convenience."""
+import contextlib
 import time
-from typing import Any, Iterator, Optional
+from typing import Any, Callable, Iterator, Optional
 
 from absl import logging
 from etils import epath
 import jax
 from jax.experimental import multihost_utils
-from jax.sharding import Mesh
 import numpy as np
 from orbax.checkpoint import type_handlers
 from orbax.checkpoint import utils
@@ -29,12 +29,38 @@ from orbax.checkpoint import utils
 PyTree = Any
 
 
+def _lock_checkpoint(checkpoint_dir: epath.Path, step: int) -> bool:
+  """Locks a checkpoint by writing a LOCKED directory."""
+  lockdir = utils.lockdir(checkpoint_dir, step)
+  try:
+    lockdir.mkdir(parents=False, exist_ok=True)
+    return True
+  except FileNotFoundError:
+    logging.info(
+        'Checkpoint step: %d was cleaned up while attempting to lock.', step
+    )
+    return False
+
+
+def _unlock_checkpoint(checkpoint_dir: epath.Path, step: int):
+  """Removes a LOCKED directory to indicate unlocking."""
+  utils.lockdir(checkpoint_dir, step).rmdir()
+
+
+def _unlock_existing_checkpoints(checkpoint_dir: epath.Path):
+  """Removes LOCKED file for all existing steps, if present."""
+  steps = utils.checkpoint_steps(checkpoint_dir)
+  for step in steps:
+    if utils.is_locked(checkpoint_dir, step):
+      _unlock_checkpoint(checkpoint_dir, step)
+
+
 def _wait_for_new_checkpoint(
     checkpoint_dir: epath.Path,
     last_checkpoint_step: Optional[int],
     seconds_to_sleep: int = 1,
     timeout: Optional[int] = None,
-) -> Optional[int]:
+) -> int:
   """Waits until a new checkpoint file is found.
 
   Args:
@@ -47,27 +73,69 @@ def _wait_for_new_checkpoint(
       process will wait indefinitely.
 
   Returns:
-    a new checkpoint step, or None if the timeout was reached.
+    a new checkpoint step, or -1 if the timeout was reached.
   """
   logging.info('Waiting for new checkpoint at %s', checkpoint_dir)
-  stop_time = time.time() + timeout if timeout is not None else None
-  while True:
-    steps = utils.checkpoint_steps(checkpoint_dir)
-    checkpoint_step = max(steps) if steps else None
-    if checkpoint_step is None or checkpoint_step == last_checkpoint_step:
-      if stop_time is not None and time.time() + seconds_to_sleep > stop_time:
-        return None
-      time.sleep(seconds_to_sleep)
-    else:
-      logging.info('Found new checkpoint step: %d', checkpoint_step)
-      return checkpoint_step
+  result = -1
+  if jax.process_index() == 0:
+    stop_time = time.time() + timeout if timeout is not None else None
+    while True:
+      steps = utils.checkpoint_steps(checkpoint_dir)
+      checkpoint_step = max(steps) if steps else None
+      if checkpoint_step is None or checkpoint_step == last_checkpoint_step:
+        if stop_time is not None and time.time() + seconds_to_sleep > stop_time:
+          break
+        time.sleep(seconds_to_sleep)
+      else:
+        if not _lock_checkpoint(checkpoint_dir, checkpoint_step):
+          continue
+        logging.info('Found new checkpoint step: %d', checkpoint_step)
+        result = checkpoint_step
+        break
+  return multihost_utils.broadcast_one_to_all(np.int32(result))
+
+
+@contextlib.contextmanager
+def wait_for_new_checkpoint(
+    checkpoint_dir: epath.Path,
+    last_checkpoint_step: Optional[int],
+    seconds_to_sleep: int = 1,
+    timeout: Optional[int] = None,
+):
+  """Waits until a new checkpoint file is found.
+
+  Automatically locks any checkpoint that is returned, and unlocks the
+  checkpoint when execution returns to this function.
+
+  Args:
+    checkpoint_dir: The directory in which checkpoints are saved.
+    last_checkpoint_step: The last checkpoint step used or `None` if we're
+      expecting a checkpoint for the first time.
+    seconds_to_sleep: The number of seconds to sleep for before looking for a
+      new checkpoint.
+    timeout: The maximum number of seconds to wait. If left as `None`, then the
+      process will wait indefinitely.
+
+  Yields:
+    a new checkpoint step, or -1 if the timeout was reached.
+  """
+  step = _wait_for_new_checkpoint(
+      checkpoint_dir, last_checkpoint_step, seconds_to_sleep, timeout
+  )
+  try:
+    yield step
+  finally:
+    # Release lock on the checkpoint step.
+    if step != -1:
+      _unlock_checkpoint(checkpoint_dir, step)
 
 
 def checkpoints_iterator(
     checkpoint_dir: epath.PathLike,
-    min_interval_secs=0,
-    timeout=None,
-    timeout_fn=None,
+    min_interval_secs: int = 0,
+    timeout: Optional[int] = None,
+    timeout_fn: Optional[Callable[[], bool]] = None,
+    unlock_existing_checkpoints: bool = True,
 ) -> Iterator[int]:
   """Continuously yield new checkpoint files as they appear.
 
@@ -110,37 +178,42 @@ def checkpoints_iterator(
     timeout_fn: Optional function to call after a timeout.  If the function
       returns True, then it means that no new checkpoints will be generated and
       the iterator will exit.  The function is called with no arguments.
+    unlock_existing_checkpoints: If True, marks any existing checkpoints as
+      "unlocked", since these were most likely created by a failure during a
+      previous run, which left some checkpoints as "locked", despite not being
+      in use.
 
   Yields:
     Integer step numbers of the latest checkpoints as they arrive.
   """
   checkpoint_dir = epath.Path(checkpoint_dir)
+  if unlock_existing_checkpoints:
+    try:
+      _unlock_existing_checkpoints(checkpoint_dir)
+    except FileNotFoundError as e:
+      logging.warning(
+          'Encountered error while unlocking existing checkpoints. Some'
+          ' checkpoints may still be locked. %s.',
+          e,
+      )
   checkpoint_step = None
   while True:
-    new_checkpoint_step = 0
-    if jax.process_index() == 0:
-      new_checkpoint_step = (
-          _wait_for_new_checkpoint(
-              checkpoint_dir, checkpoint_step, timeout=timeout
-          )
-          or -1
-      )
-    new_checkpoint_step = multihost_utils.broadcast_one_to_all(
-        np.int32(new_checkpoint_step)
-    )
-    if new_checkpoint_step == -1:
-      if not timeout_fn:
-        logging.info('Timed-out waiting for a checkpoint.')
-        return
-      if timeout_fn():
-        # The timeout_fn indicated that we are truly done.
-        return
-      else:
-        # The timeout_fn indicated that more checkpoints may come.
-        continue
-    start = time.time()
-    checkpoint_step = new_checkpoint_step
-    yield checkpoint_step
+    with wait_for_new_checkpoint(
+        checkpoint_dir, checkpoint_step, timeout=timeout
+    ) as new_checkpoint_step:
+      if new_checkpoint_step == -1:
+        if not timeout_fn:
+          logging.info('Timed-out waiting for a checkpoint.')
+          return
+        if timeout_fn():
+          # The timeout_fn indicated that we are truly done.
+          return
+        else:
+          # The timeout_fn indicated that more checkpoints may come.
+          continue
+      start = time.time()
+      checkpoint_step = new_checkpoint_step
+      yield checkpoint_step
     time_to_next_eval = start + min_interval_secs - time.time()
     if time_to_next_eval > 0:
       time.sleep(time_to_next_eval)
@@ -148,7 +221,7 @@ def checkpoints_iterator(
 
 # TODO(b/274813763): Remove this function when no longer depended on by Flax.
 def restore_args_from_target(
-    mesh: Mesh, target: PyTree, axes_tree: PyTree
+    mesh: jax.sharding.Mesh, target: PyTree, axes_tree: PyTree
 ) -> PyTree:
   """DEPRECATED, DO NOT USE.
 
@@ -184,7 +257,7 @@ def restore_args_from_target(
   ignored.
 
   Args:
-    mesh: Mesh to shard arrays with.
+    mesh: jax.sharding.Mesh to shard arrays with.
     target: The returned value will match the structure of `target`, will be
       used to set the desired dtype and restoration shape.
     axes_tree: A PyTree matching `target` which will be used to set the